move MyClusterRenderer.java to utils
